# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InteriorHub Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalEngineers: number;
  pendingApprovals: number;
  totalProjects: number;
  totalCategories: number;
  approvedEngineers: number;
  disabledEngineers: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  engineer_signup: 'engineer_signup',
  project_added: 'project_added',
  engineer_approved: 'engineer_approved',
  engineer_disabled: 'engineer_disabled',
  contact_request: 'contact_request',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  timestamp: string;
  /** @nullable */
  relatedId?: number | null;
  /** @nullable */
  relatedName?: string | null;
}

export type EngineerStatus = typeof EngineerStatus[keyof typeof EngineerStatus];


export const EngineerStatus = {
  pending: 'pending',
  approved: 'approved',
  disabled: 'disabled',
} as const;

export interface Engineer {
  id: number;
  fullName: string;
  city: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profilePhoto?: string | null;
  status: EngineerStatus;
  /** @nullable */
  trialEndDate?: string | null;
  projectsCount: number;
  createdAt: string;
}

export interface EngineerListResponse {
  engineers: Engineer[];
  total: number;
  page: number;
  totalPages: number;
}

export type EngineerStatusUpdateStatus = typeof EngineerStatusUpdateStatus[keyof typeof EngineerStatusUpdateStatus];


export const EngineerStatusUpdateStatus = {
  pending: 'pending',
  approved: 'approved',
  disabled: 'disabled',
} as const;

export interface EngineerStatusUpdate {
  status: EngineerStatusUpdateStatus;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  projectsCount: number;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  slug?: string;
}

export interface DeleteResult {
  success: boolean;
}

export interface Project {
  id: number;
  title: string;
  /** @nullable */
  coverImage?: string | null;
  categoryId: number;
  categoryName: string;
  engineerId: number;
  engineerName: string;
  createdAt: string;
}

export interface ProjectDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImage?: string | null;
  images: string[];
  categoryId: number;
  categoryName: string;
  engineerId: number;
  engineerName: string;
  engineerEmail: string;
  createdAt: string;
}

export interface ProjectListResponse {
  projects: Project[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ContactRequest {
  id: number;
  engineerId: number;
  engineerName: string;
  engineerEmail: string;
  clientName: string;
  clientEmail: string;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}

export type ListEngineersParams = {
search?: string;
city?: string;
status?: ListEngineersStatus;
page?: number;
limit?: number;
};

export type ListEngineersStatus = typeof ListEngineersStatus[keyof typeof ListEngineersStatus];


export const ListEngineersStatus = {
  pending: 'pending',
  approved: 'approved',
  disabled: 'disabled',
} as const;

export type ListProjectsParams = {
categoryId?: number;
engineerId?: number;
page?: number;
limit?: number;
};

export type ListContactRequestsParams = {
engineerId?: number;
};