# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InteriorHub Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalEngineers: number;
  pendingApprovals: number;
  totalProjects: number;
  totalCategories: number;
  approvedEngineers: number;
  disabledEngineers: number;
}