# lib/api-zod/src/generated/types/engineer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InteriorHub Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { EngineerStatus } from './engineerStatus';

export interface Engineer {
  id: number;
  fullName: string;
  city: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profilePhoto?: string | null;
  status: EngineerStatus;
  /** @nullable */
  trialEndDate?: string | null;
  projectsCount: number;
  createdAt: string;
}