# lib/api-zod/src/generated/types/engineerStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InteriorHub Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { EngineerStatusUpdateStatus } from './engineerStatusUpdateStatus';

export interface EngineerStatusUpdate {
  status: EngineerStatusUpdateStatus;
}