# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InteriorHub Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverImage?: string | null;
  images: string[];
  categoryId: number;
  categoryName: string;
  engineerId: number;
  engineerName: string;
  engineerEmail: string;
  createdAt: string;
}